refactor(workbench): extract composer keyboard FSM into useComposerKeyDown
The composer keydown handler was a 228-line inline ref assignment inside
AppWorkbench covering four concerns: @-file menu nav, slash/plus palette
nav, prompt-history browse (picker + CLI-like arrows), and the Enter
send-vs-steer resolution plus Escape unwind.

Moves the FSM verbatim into src/hooks/useComposerKeyDown.ts behind a
typed ComposerKeyDownCtx bag. The hook keeps the existing ref-rebind
pattern (ref.current reassigned each render), so the handler still reads
the freshest snapshot — identical semantics, no behavior change, and the
handler stays off the memo-dep graph as before.

AppWorkbench: 13634 -> 13438 lines.
